refactor: Разгружен слой app
