+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next"
 import { Geist, Geist_Mono } from "next/font/google"
 import "./globals.css"
-import { StoreProvider } from "@/app/providers/store-provider"
+import { StoreProvider } from "@/shared/providers"
 import { Header } from "@/widgets/header/header"
 import { Alert } from "@/features/alert/alert"
