+++ src/shared/lib/redux/useAppDispatch.ts
@@ -1,4 +1,4 @@
-import { AppDispatch } from "@/app/store/store"
+import { AppDispatch } from "@/shared/store/store"
 import { useDispatch } from "react-redux"
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>()
