+++ src/shared/lib/redux/useAppSelector.ts
@@ -1,4 +1,4 @@
 import { useSelector } from "react-redux"
-import { AppRootState } from "@/app/store/store"
+import { AppRootState } from "@/shared/store/store"
 
 export const useAppSelector = useSelector.withTypes<AppRootState>()
