+++ src/shared/providers/StoreProvider.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { Provider } from "react-redux"
-import { store } from "@/app/store/store"
+import { store } from "@/shared/store/store"
 
 export function StoreProvider({ children }: { children: React.ReactNode }) {
   return <Provider store={store}>{children}</Provider>
