+++ src/shared/providers/index.ts
@@ -1,0 +1,1 @@
+export { StoreProvider } from "./StoreProvider"
